+++ lib/parse/model/object.rb
@@ -346,10 +346,11 @@ class Array
   # if it constains the proper fields. Non convertible objects will be removed
   # If the className is not contained or known, you can pass a table name as an argument
   def parse_objects(table = nil)
+    f = "className".freeze
     map do |m|
       next m if m.is_a?(Parse::Pointer)
-      if m.is_a?(Hash) && (m["className"] || m[:className] || table)
-        next Parse::Object.build m, (m["className"] || m[:className] || table)
+      if m.is_a?(Hash) && (m[f] || m[:className] || table)
+        next Parse::Object.build m, (m[f] || m[:className] || table)
       end
       nil
     end.compact